+++ www/views/Extjs3/javascript/go/links/CreateLinkWindow.js
@@ -3,6 +3,13 @@ go.links.CreateLinkWindow = Ext.extend(go.Window, {
 	entityId: null,
 	modal: true,
 	stateId: "go-create-link-windows",
+	
+	/**
+	 * Provide the entities to show in the list here
+	 * When not provided, the list will show all entities
+	 */
+	entities : null,
+	
 	search: function (v) {
 		
 		var filter = {};
@@ -61,7 +68,8 @@ go.links.CreateLinkWindow = Ext.extend(go.Window, {
 		this.entityGrid = new go.links.EntityGrid({
 			width: dp(200),
 			region:"west",
-			savedSelection: "link"
+			savedSelection: "link",
+			entities:this.entities
 		});
 		
 		this.entityGrid.getSelectionModel().on('selectionchange', function (sm) {